refactor: ♻️ (`main.py`) 共有状態を持つサブスレッドを並列して実行できるようにした。
顔認証関連のコードを、`main.py`から`face.py`に移動させた。
